requirements.txt
face-recognition==1.3.0
numpy==1.24.3
opencv-python==4.8.0.76
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pymongo==4.5.0
langchain==0.0.267
google-generativeai==0.3.1
langchain-google-genai==0.0.5
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.3.0
loguru==0.7.0
requests==2.31.0
pillow==10.0.0
